Restructure workflow validation modules
