+++ cg/services/order_validation_service/workflows/tomte/models/case.py
@@ -3,7 +3,7 @@ from pydantic import AfterValidator, model_validator
 from cg.constants import DataDelivery, GenePanelMasterList
 from cg.models.orders.orderform_schema import OrderCase
 from cg.services.order_validation_service.models.case_validators import validate_subject_id
-from cg.services.order_validation_service.models.tomte.tomte_case_validators import (
+from cg.services.order_validation_service.workflows.tomte.validators.case_validators import (
     validate_tomte_delivery_type,
 )
