+++ cg/services/order_validation_service/workflows/tomte/models/sample.py
@@ -2,7 +2,7 @@ from pydantic import Field, model_validator
 
 from cg.constants.constants import GenomeVersion
 from cg.models.orders.sample_base import NAME_PATTERN, ControlEnum, SexEnum, StatusEnum
-from cg.services.validation_service.constants import TissueBlockEnum
+from cg.services.order_validation_service.constants import TissueBlockEnum
 from cg.services.order_validation_service.models.order_sample import OrderSample
 from cg.services.order_validation_service.models.sample_validators import (
     validate_ffpe_source,
